fix: clean up live room implementation
